Sum of N natural Numbers using recursion

+++ Training/SumofN.cpp
@@ -1,13 +1,12 @@
 #include<iostream>
 using namespace std;
 int main(){
-    int n,sum=0;
+    long long n,sum=0;
     cout<<"Enter The Value of N : ";
     cin>>n;
-    for (int i =0; i<=n ; i++){
+    for (long long i =0; i<=n ; i++){
         sum += i;
     }
     cout<<"Sum is : "<<sum<<endl;
     return 0;
 }
-
